+++ public/config.js
@@ -186,7 +186,6 @@
     logoDisabled: false,
     configNsId: null,
     configBaseURL: '',
-    tableMargin: [16, 16, 16, 16],
     theme: theme,
     layout: layout,
     navProps: layoutMap[layout].navProps,